requirements.txt
python-hcl2
openpyxl